Backend/requirements.txt
# Backend/requirements.txt
# Core backend dependencies for Liberty AI Trade.
# Restored to match the immutable structure in README_UPDATER.txt.

fastapi
uvicorn[standard]
pydantic
pydantic-settings
aiohttp
requests
websockets
python-telegram-bot
PyJWT
pytest
psutil
